Update rank classification passthrough keys to work when there are an unknown amout of targets for a given example.
With some datasets, like ReCoRD, support rank classification with a different number of possible labels for each example (in ReCoRD this is the provided list of entities). In this case, the length of the target vector is not known at graph construction time. When this is the case, the current method of passing extra keys through the rank classification preprocessor (multiplying a list of the extra value by the `len` of the target and converting that to a tensor) doesn't work.

This change updates the rank classification preprocessor to use `tf.tile` and `tf.shape` to create the copies of the extra values dynamically that match the shape of the targets.

This also updates to the rank classification preprocessor test to include an extra value that is a list, and to use a dataset generator, allowing us to have a dataset where the shape of the target is unkown.

PiperOrigin-RevId: 440452343